refactor tests
I move some helper code closer to the only test that uses it.
